[#6] Extract `LoginForm` to separate file
