+++ lib/screens/login/login_form.dart
@@ -1,0 +1,83 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:survey_flutter/screens/widgets/form_field_decoration.dart';
+import 'package:survey_flutter/utils/keyboard_manager.dart';
+import 'package:survey_flutter/theme/constant.dart';
+
+class LoginForm extends StatefulWidget {
+  const LoginForm({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _LoginFormState();
+}
+
+class _LoginFormState extends State<LoginForm> {
+  AppLocalizations get _localizations => AppLocalizations.of(context)!;
+  TextTheme get _textTheme => Theme.of(context).textTheme;
+
+  TextFormField get _emailTextField => TextFormField(
+        keyboardType: TextInputType.emailAddress,
+        autocorrect: false,
+        decoration: FormFieldDecoration(
+          hint: _localizations.emailInputHint,
+          hintTextStyle: _textTheme.bodyLarge,
+        ),
+        style: _textTheme.bodyLarge,
+      );
+
+  TextFormField get _passwordTextField => TextFormField(
+        keyboardType: TextInputType.emailAddress,
+        obscureText: true,
+        obscuringCharacter: "●",
+        decoration: FormFieldDecoration(
+          hint: _localizations.passwordInputHint,
+          hintTextStyle: _textTheme.bodyLarge,
+        ),
+        style: _textTheme.bodyLarge,
+      );
+
+  ElevatedButton get _loginButton => ElevatedButton(
+        style: ButtonStyle(
+          textStyle:
+              MaterialStateProperty.all(Theme.of(context).textTheme.labelLarge),
+          backgroundColor: MaterialStateProperty.all(Colors.white),
+          foregroundColor: MaterialStateProperty.all(Colors.black),
+          overlayColor: MaterialStateProperty.all(Colors.black12),
+          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+            RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(Metrics.defaultBorderRadius),
+            ),
+          ),
+        ),
+        child: Text(_localizations.loginButton),
+        onPressed: () {
+          KeyboardManager.dismiss(context);
+          // TODO - implement login
+        },
+      );
+
+  @override
+  Widget build(BuildContext context) {
+    const fieldSpacing = 20.0;
+    return Form(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          _emailTextField,
+          const SizedBox(
+            height: fieldSpacing,
+          ),
+          _passwordTextField,
+          const SizedBox(
+            height: fieldSpacing,
+          ),
+          SizedBox(
+            height: 56,
+            width: double.infinity,
+            child: _loginButton,
+          )
+        ],
+      ),
+    );
+  }
+}
